[refactor] 잔액 구조 변경 및 repo 분리
