+++ src/main/java/kr/hhplus/be/server/apidocs/BalanceApiDocs.java
@@ -5,9 +5,9 @@ import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import kr.hhplus.be.server.common.ApiResponse;
-import kr.hhplus.be.server.domain.balance.controller.dto.ChargeBalanceRequest;
-import kr.hhplus.be.server.domain.balance.controller.dto.ChargeBalanceResponse;
-import kr.hhplus.be.server.domain.balance.controller.dto.ViewBalanceResponse;
+import kr.hhplus.be.server.domain.balance.presenter.controller.dto.ChargeBalanceRequest;
+import kr.hhplus.be.server.domain.balance.presenter.controller.dto.ChargeBalanceResponse;
+import kr.hhplus.be.server.domain.balance.presenter.controller.dto.ViewBalanceResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
