+++ src/main/java/kr/hhplus/be/server/domain/balance/application/service/BalanceHistoryService.java
@@ -1,7 +1,7 @@
-package kr.hhplus.be.server.domain.balance.service;
+package kr.hhplus.be.server.domain.balance.application.service;
 
-import kr.hhplus.be.server.domain.balance.entity.BalanceHistory;
-import kr.hhplus.be.server.domain.balance.repository.BalanceHistoryRepository;
+import kr.hhplus.be.server.domain.balance.domain.entity.BalanceHistory;
+import kr.hhplus.be.server.domain.balance.domain.repository.BalanceHistoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
