+++ src/main/java/kr/hhplus/be/server/domain/balance/application/service/dto/ChargeBalanceCommand.java
@@ -1,6 +1,6 @@
-package kr.hhplus.be.server.domain.balance.service.dto;
+package kr.hhplus.be.server.domain.balance.application.service.dto;
 
-import kr.hhplus.be.server.domain.balance.controller.dto.ChargeBalanceRequest;
+import kr.hhplus.be.server.domain.balance.presenter.controller.dto.ChargeBalanceRequest;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
