+++ src/main/java/kr/hhplus/be/server/domain/balance/domain/repository/BalanceHistoryRepository.java
@@ -1,6 +1,6 @@
-package kr.hhplus.be.server.domain.balance.repository;
+package kr.hhplus.be.server.domain.balance.domain.repository;
 
-import kr.hhplus.be.server.domain.balance.entity.BalanceHistory;
+import kr.hhplus.be.server.domain.balance.domain.entity.BalanceHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BalanceHistoryRepository extends JpaRepository<BalanceHistory, Long> {
