+++ src/main/java/kr/hhplus/be/server/domain/balance/domain/repository/BalanceRepository.java
@@ -1,0 +1,14 @@
+package kr.hhplus.be.server.domain.balance.domain.repository;
+
+import kr.hhplus.be.server.domain.balance.domain.entity.Balance;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface BalanceRepository {
+
+    Optional<Balance> findById(String userId);
+
+    Balance save(Balance balance);
+}
