+++ src/main/java/kr/hhplus/be/server/domain/balance/infra/BalanceJpaRepository.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.domain.balance.infra;
+
+import kr.hhplus.be.server.domain.balance.domain.entity.Balance;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BalanceJpaRepository extends JpaRepository<Balance, String> {
+}
