+++ src/main/java/kr/hhplus/be/server/domain/balance/presenter/controller/BalanceController.java
@@ -1,16 +1,16 @@
-package kr.hhplus.be.server.domain.balance.controller;
+package kr.hhplus.be.server.domain.balance.presenter.controller;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
 import kr.hhplus.be.server.apidocs.BalanceApiDocs;
 import kr.hhplus.be.server.common.ApiResponse;
-import kr.hhplus.be.server.domain.balance.controller.dto.ChargeBalanceRequest;
-import kr.hhplus.be.server.domain.balance.controller.dto.ChargeBalanceResponse;
-import kr.hhplus.be.server.domain.balance.controller.dto.ViewBalanceResponse;
-import kr.hhplus.be.server.domain.balance.service.BalanceService;
-import kr.hhplus.be.server.domain.balance.service.dto.ChargeBalanceCommand;
-import kr.hhplus.be.server.domain.balance.service.dto.ChargeBalanceResult;
-import kr.hhplus.be.server.domain.balance.service.dto.ViewBalanceCommand;
-import kr.hhplus.be.server.domain.balance.service.dto.ViewBalanceResult;
+import kr.hhplus.be.server.domain.balance.presenter.controller.dto.ChargeBalanceRequest;
+import kr.hhplus.be.server.domain.balance.presenter.controller.dto.ChargeBalanceResponse;
+import kr.hhplus.be.server.domain.balance.presenter.controller.dto.ViewBalanceResponse;
+import kr.hhplus.be.server.domain.balance.application.service.BalanceService;
+import kr.hhplus.be.server.domain.balance.application.service.dto.ChargeBalanceCommand;
+import kr.hhplus.be.server.domain.balance.application.service.dto.ChargeBalanceResult;
+import kr.hhplus.be.server.domain.balance.application.service.dto.ViewBalanceCommand;
+import kr.hhplus.be.server.domain.balance.application.service.dto.ViewBalanceResult;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
